+++ src/GRA.Controllers/MissionControl/ParticipantsController.cs
@@ -432,7 +432,7 @@ namespace GRA.Controllers.MissionControl
                     viewModel.SchoolDistrictList = new SelectList(districtList.ToList(), "Id", "Name");
                 }
 
-                if (UserHasPermission(Permission.EditParticipantUsernames) 
+                if (UserHasPermission(Permission.EditParticipantUsernames)
                     && !string.IsNullOrWhiteSpace(user.Username))
                 {
                     viewModel.CanEditUsername = true;
@@ -1384,7 +1384,14 @@ namespace GRA.Controllers.MissionControl
         [Authorize(Policy = Policy.ViewUserPrizes)]
         public async Task<IActionResult> RedeemWinner(int prizeWinnerId, int userId, int page = 1)
         {
-            await _prizeWinnerService.RedeemPrizeAsync(prizeWinnerId);
+            try
+            {
+                await _prizeWinnerService.RedeemPrizeAsync(prizeWinnerId);
+            }
+            catch (GraException gex)
+            {
+                ShowAlertWarning("Unable to redeem prize: ", gex);
+            }
             return RedirectToAction("Prizes", new { id = userId, page = page });
         }
 
@@ -1392,7 +1399,14 @@ namespace GRA.Controllers.MissionControl
         [Authorize(Policy = Policy.ViewUserPrizes)]
         public async Task<IActionResult> UndoRedemption(int prizeWinnerId, int userId, int page = 1)
         {
-            await _prizeWinnerService.UndoRedemptionAsync(prizeWinnerId);
+            try
+            {
+                await _prizeWinnerService.UndoRedemptionAsync(prizeWinnerId);
+            }
+            catch (GraException gex)
+            {
+                ShowAlertWarning("Unable to undo redemption: ", gex);
+            }
             return RedirectToAction("Prizes", new { id = userId, page = page });
         }
         #endregion