+++ CHANGELOG.md
@@ -39,6 +39,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - Fix trigger badge creation exception
 - Fix removing drawing winners
 - Fix service broadcast mail permission
+- Fix prize redemption exceptions not being caught
 
 ## [4.0.0-alpha5] 2017-05-01
 ### Added